Extract ArticleModel interface into own file
