+++ packages/frontend/amp/components/topMeta/TopMeta.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { ArticleModel } from '@frontend/amp/pages/Article';
+import { ArticleModel } from '@frontend/amp/types/ArticleModel';
 import { TopMetaNews } from '@frontend/amp/components/topMeta/TopMetaNews';
 import { TopMetaOpinion } from '@frontend/amp/components/topMeta/TopMetaOpinion';
 import { TopMetaPaidContent } from '@frontend/amp/components/topMeta/TopMetaPaidContent';
