+++ packages/frontend/amp/pages/Article.tsx
@@ -13,43 +13,11 @@ import { filterForTagsOfType } from '@frontend/amp/lib/tag-utils';
 import { AdUserSync } from '@root/packages/frontend/amp/components/AdUserSync';
 import { getPillar } from '@frontend/lib/pillars';
 import { palette } from '@guardian/src-foundations';
+import { ArticleModel } from '@frontend/amp/types/ArticleModel';
 
 const backgroundColour = css`
     background-color: ${palette.neutral[97]};
 `;
-
-export interface ArticleModel {
-    headline: string;
-    standfirst: string;
-    webTitle: string;
-    mainMediaElements: CAPIElement[];
-    keyEvents: Block[]; // liveblog-specific
-    pagination?: Pagination;
-    blocks: Block[];
-    author: AuthorType;
-    webPublicationDate: string;
-    webPublicationDateDisplay: string;
-    pageId: string;
-    pillar: Pillar;
-    sectionLabel?: string;
-    sectionUrl?: string;
-    sectionName?: string;
-    tags: TagType[];
-    subMetaSectionLinks: SimpleLinkType[];
-    subMetaKeywordLinks: SimpleLinkType[];
-    webURL: string;
-    shouldHideAds: boolean;
-    guardianBaseURL: string;
-    hasRelated: boolean;
-    hasStoryPackage: boolean;
-    isCommentable: boolean;
-    editionId: Edition;
-    contentType: string;
-    commercialProperties: CommercialProperties;
-    isImmersive: boolean;
-    starRating?: number;
-    designType: DesignType;
-}
 
 const Body: React.SFC<{
     data: ArticleModel;
