+++ packages/frontend/amp/types/ArticleModel.tsx
@@ -1,0 +1,32 @@
+export interface ArticleModel {
+    headline: string;
+    standfirst: string;
+    webTitle: string;
+    mainMediaElements: CAPIElement[];
+    keyEvents: Block[]; // liveblog-specific
+    pagination?: Pagination;
+    blocks: Block[];
+    author: AuthorType;
+    webPublicationDate: string;
+    webPublicationDateDisplay: string;
+    pageId: string;
+    pillar: Pillar;
+    sectionLabel?: string;
+    sectionUrl?: string;
+    sectionName?: string;
+    tags: TagType[];
+    subMetaSectionLinks: SimpleLinkType[];
+    subMetaKeywordLinks: SimpleLinkType[];
+    webURL: string;
+    shouldHideAds: boolean;
+    guardianBaseURL: string;
+    hasRelated: boolean;
+    hasStoryPackage: boolean;
+    isCommentable: boolean;
+    editionId: Edition;
+    contentType: string;
+    commercialProperties: CommercialProperties;
+    isImmersive: boolean;
+    starRating?: number;
+    designType: DesignType;
+}
